+++ src/main/java/workdone/exception/MissingCommandDetailException.java
